requirements.txt
fastapi 
uvicorn 
sentence-transformers 
python-dotenv 
requests 
numpy
sqlalchemy
alembic
psycopg2-binary